Renamed `block_table.h` to `block_table.hpp`
